+++ android/src/main/kotlin/com/camerakit/camerakit_flutter/Configuration.kt
@@ -1,3 +1,5 @@
+package com.camerakit.camerakit_flutter
+
 //
 //  Configuration.kt
 //  com.camerakit.camerakit_flutter
@@ -5,9 +7,8 @@
 //  Copyright © 2023 DevCrew I/O
 //
 class Configuration private constructor(
-    val cameraKitApiToken: String,
-    val groupIds: List<String>,
-    val lensId: String
+    val apiToken: String,
+    var isHideCloseButton: Boolean
 ) {
 
     companion object {
@@ -19,9 +20,8 @@ class Configuration private constructor(
 
         fun createFromMap(arguments: Map<String, Any>): Configuration {
             instance = Configuration(
-                groupIds = arguments["groupIds"] as? List<String> ?: emptyList(),
-                cameraKitApiToken = arguments["token"] as? String ?: "",
-                lensId = arguments["lensId"] as? String ?: "",
+                apiToken = arguments["token"] as? String ?: "",
+                isHideCloseButton = false,
             )
             return instance!!
         }