+++ example/lib/constants.dart
@@ -4,9 +4,6 @@ class Constants {
   /// List of group IDs for Camera Kit
   static const List<String> groupIdList = ['your-group-ids']; // TODO: Fill group IDs here
 
-  /// Optional: if you want to get single len from a group you can set lens id here otherwise set empty string
-  static const cameraKitLensId = ''; // TODO: Fill lens ID here
-
   /// The API token for Camera Kit in the staging environment
   static const cameraKitApiToken = 'your-api-token'; // TODO fill api token staging or production here
 }