+++ example/lib/main.dart
@@ -1,4 +1,3 @@
-import 'package:camerakit_flutter/configuration_camerakit.dart';
 import 'package:camerakit_flutter_example/media_result_screen.dart';
 import 'package:camerakit_flutter_example/lens_list_screen.dart';
 import 'package:flutter/foundation.dart';
@@ -35,11 +34,8 @@ class _MyAppState extends State<MyApp> implements CameraKitFlutterEvents {
   @override
   void initState() {
     super.initState();
-    final config = Configuration(
-      Constants.cameraKitApiToken,
-      Constants.groupIdList);
 
-    _cameraKitFlutterImpl.setCredentials(config);
+    _cameraKitFlutterImpl.setCredentials(apiToken: Constants.cameraKitApiToken);
   }
 
   // Platform messages are asynchronous, so we initialize in an async method.
@@ -47,7 +43,8 @@ class _MyAppState extends State<MyApp> implements CameraKitFlutterEvents {
     // Platform messages may fail, so we use a try/catch PlatformException.
     // We also handle the message potentially returning null.
     try {
-      await _cameraKitFlutterImpl.openCameraKit();
+      await _cameraKitFlutterImpl.openCameraKit(
+          groupIds: Constants.groupIdList, isHideCloseButton: false);
     } on PlatformException {
       if (kDebugMode) {
         print("Failed to open camera kit");
@@ -60,7 +57,7 @@ class _MyAppState extends State<MyApp> implements CameraKitFlutterEvents {
     // Platform messages may fail, so we use a try/catch PlatformException.
     // We also handle the message potentially returning null.
     try {
-      _cameraKitFlutterImpl.getGroupLenses(Constants.groupIdList);
+      _cameraKitFlutterImpl.getGroupLenses(groupIds: Constants.groupIdList);
     } on PlatformException {
       if (kDebugMode) {
         print("Failed to open camera kit");
@@ -116,7 +113,15 @@ class _MyAppState extends State<MyApp> implements CameraKitFlutterEvents {
   void receivedLenses(List<Lens> lensList) async {
     isLensListPressed = false;
     setState(() {});
-    await Navigator.of(context).push(MaterialPageRoute(
-        builder: (context) => LensListView(lensList: lensList)));
+    final result = await Navigator.of(context).push(MaterialPageRoute(
+            builder: (context) => LensListView(lensList: lensList)))
+        as Map<String, dynamic>?;
+    final lensId = result?['lensId'] as String?;
+    final groupId = result?['groupId'] as String?;
+
+    if ((lensId?.isNotEmpty ?? false) && (groupId?.isNotEmpty ?? false)) {
+      _cameraKitFlutterImpl.openCameraKitWithSingleLens(
+          lensId: lensId!, groupId: groupId!, isHideCloseButton: false);
+    }
   }
 }