+++ test/camerakit_flutter_method_channel_test.dart
@@ -22,8 +22,4 @@ void main() {
     TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
         .setMockMethodCallHandler(channel, null);
   });
-
-  test('getPlatformVersion', () async {
-    expect(await platform.openCameraKit(), '42');
-  });
 }