+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## 0.0.7
+### Fixed
+* `Low Quality` output image (iOS).
+* Release `Camera` and `Microphone` access when closing the camera screen (iOS).
+
+### Added
+* `Close Button` on camera screen, Default it is visible but can hide it with `isHideCloseButton` to true.
+* Open the camera screen with a single lens when the user taps on a lens list.
+
+### Updated
+* Package implementation guide.
+* Example Project
+* README.md
+
 ## 0.0.6
 * Make cameraKitLensId optional
 * Snap CameraKit Upgraded to 1.27.0